pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sac-mujoco"
version = "0.1.0"
description = "Research-grade Soft Actor-Critic training pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "gymnasium>=0.29",
  "numpy>=1.24",
  "torch>=2.1",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
